+++ app/src/main/java/com/techcourse/controller/annotation_controller/RegisterAnnotationController.java
@@ -1,4 +1,4 @@
-package com.techcourse.controller;
+package com.techcourse.controller.annotation_controller;
 
 import com.techcourse.domain.User;
 import com.techcourse.repository.InMemoryUserRepository;
