+++ app/src/main/java/com/techcourse/controller/manual_controller/LogoutController.java
@@ -1,13 +1,16 @@
-package com.techcourse.controller;
+package com.techcourse.controller.manual_controller;
 
+import com.techcourse.controller.UserSession;
 import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpServletResponse;
 import webmvc.org.springframework.web.servlet.mvc.asis.Controller;
 
-public class RegisterViewController implements Controller {
+public class LogoutController implements Controller {
 
     @Override
     public String execute(final HttpServletRequest req, final HttpServletResponse res) throws Exception {
-        return "/register.jsp";
+        final var session = req.getSession();
+        session.removeAttribute(UserSession.SESSION_KEY);
+        return "redirect:/";
     }
 }
