+++ app/src/main/java/com/techcourse/controller/manual_controller/RegisterViewController.java
@@ -1,15 +1,13 @@
-package com.techcourse.controller;
+package com.techcourse.controller.manual_controller;
 
 import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpServletResponse;
 import webmvc.org.springframework.web.servlet.mvc.asis.Controller;
 
-public class LogoutController implements Controller {
+public class RegisterViewController implements Controller {
 
     @Override
     public String execute(final HttpServletRequest req, final HttpServletResponse res) throws Exception {
-        final var session = req.getSession();
-        session.removeAttribute(UserSession.SESSION_KEY);
-        return "redirect:/";
+        return "/register.jsp";
     }
 }
